+++ examples/mujoco/mujoco_ddpg.py
@@ -6,7 +6,7 @@ import pprint
 
 import numpy as np
 import torch
-from mujoco_env import make_mujoco_env
+from gymnasium.spaces import Box
 from sensai.util import logging
 
 from tianshou.algorithm import DDPG
@@ -14,6 +14,7 @@ from tianshou.algorithm.algorithm_base import Algorithm
 from tianshou.algorithm.modelfree.ddpg import ContinuousDeterministicPolicy
 from tianshou.algorithm.optim import AdamOptimizerFactory
 from tianshou.data import Collector, CollectStats, ReplayBuffer, VectorReplayBuffer
+from tianshou.env.mujoco import make_mujoco_env
 from tianshou.exploration import GaussianNoise
 from tianshou.highlevel.logger import LoggerFactoryDefault
 from tianshou.trainer import OffPolicyTrainerParams
@@ -68,8 +69,10 @@ def main(
         num_test_envs,
         obs_norm=False,
     )
-    state_shape = env.observation_space.shape or env.observation_space.n
-    action_shape = env.action_space.shape or env.action_space.n
+    assert isinstance(env.observation_space, Box)
+    assert isinstance(env.action_space, Box)
+    state_shape = env.observation_space.shape
+    action_shape = env.action_space.shape
     max_action = env.action_space.high[0]
     exploration_noise = exploration_noise * max_action
     log.info(f"Observations shape: {state_shape}")